requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.103.1
flake8==6.1.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
mccabe==0.7.0
mysql-connector-python==8.1.0
protobuf==4.21.12
pycodestyle==2.11.0
pydantic==2.3.0
pydantic_core==2.6.3
pyflakes==3.1.0
python-decouple==3.8
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
tabulate==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
